+++ R/worker.R
@@ -330,6 +330,10 @@ append_log = function(log = NULL, stage = NA_character_, class = NA_character_, 
   }
 
   if (length(msg)) {
+    pwalk(list(stage, class, msg), function(s, c, m) {
+      if (c == "error") lg$error("%s: %s", s, m)
+      if (c == "warning") lg$warn("%s: %s", s, m)
+    })
     log = rbindlist(list(log, data.table(stage = stage, class = class, msg = msg)), use.names = TRUE)
   }