+++ R/LearnerClassifDebug.R
@@ -24,7 +24,7 @@
 #'    \item{warning_train:}{Probability to signal a warning during train.}
 #'    \item{x:}{Numeric tuning parameter. Has no effect.}
 #'    \item{iter:}{Integer parameter for testing hotstarting.}
-#'    \item{count_marshaling:}{If `TRUE`, `marshal_model` will increase the `marshal_count` by 1. This value is initialized to `FALSE`.}
+#'    \item{count_marshaling:}{If `TRUE`, `marshal_model` will increase the `marshal_count` by 1 each time it is called. The default is `FALSE`.}
 #' }
 #' Note that segfaults may not be triggered reliably on your operating system.
 #' Also note that if they work as intended, they will tear down your R session immediately!