+++ R/worker.R
@@ -1,4 +1,4 @@
-learner_train = function(learner, task, train_row_ids = NULL, test_row_ids = NULL, mode = "train", unmarshal = TRUE) {
+learner_train = function(learner, task, train_row_ids = NULL, test_row_ids = NULL, mode = "train") {
   # This wrapper calls learner$.train, and additionally performs some basic
   # checks that the training was successful.
   # Exceptions here are possibly encapsulated, so that they get captured
@@ -71,8 +71,6 @@ learner_train = function(learner, task, train_row_ids = NULL, test_row_ids = NUL
 
   log = append_log(NULL, "train", result$log$class, result$log$msg)
   train_time = result$elapsed
-
-  # unmarshal_model does nothing if the model was not marshaled
 
   learner$state = set_class(insert_named(learner$state, list(
     model = result$result,
@@ -223,7 +221,7 @@ learner_predict = function(learner, task, row_ids = NULL) {
 }
 
 
-workhorse = function(iteration, task, learner, resampling, param_values = NULL, lgr_threshold, store_models = FALSE, pb = NULL, mode = "train", is_sequential = TRUE) {
+workhorse = function(iteration, task, learner, resampling, param_values = NULL, lgr_threshold, store_models = FALSE, pb = NULL, mode = "train", is_sequential = TRUE, unmarshal = TRUE) {
   if (!is.null(pb)) {
     pb(sprintf("%s|%s|i:%i", task$id, learner$id, iteration))
   }
@@ -259,17 +257,17 @@ workhorse = function(iteration, task, learner, resampling, param_values = NULL, 
     learner$param_set$set_values(.values = param_values)
   }
   learner_hash = learner$hash
-  learner = learner_train(learner, task, sets[["train"]], sets[["test"]], mode = mode, unmarshal = FALSE)
+  learner = learner_train(learner, task, sets[["train"]], sets[["test"]], mode = mode)
 
   # The model is in marshaled form in case we did callr encapsulation and the model needed marshaling.
-  # But for the predict step we need it unmarshaled.
-  # In case we have a marshaled model and are parallelizing, we avoid to marshal again after unmarshaling
-  # the model for the predict step.
-  # We do this by temporarily keep the marshaled and unmarshaled model in RAM (the marshaled model model of course only
-  # if store_models is TRUE)
-  model_marshaled_or_null = unmarshal_maybe_after_train(
-    learner = learner, store_models = store_models, is_sequential = is_sequential
+  # For the predict step we need it unmarshaled.
+  # If the model is later sent back to main process (when parallelizing) and needs marshaling and the model is already
+  # in marshaled form, we preserve the marshaled model.
+  # This means, we temporarily keep both the marshaled and unmarshaled model in RAM
+  model_marshaled_or_null = get_marshaled_if_needed(
+    learner = learner, store_models = store_models, is_sequential = is_sequential, unmarshal = unmarshal
   )
+  learner$model = unmarshal_model(model = learner$model, inplace = FALSE)
 
   # predict for each set
   sets = sets[learner$predict_sets]
@@ -279,11 +277,11 @@ workhorse = function(iteration, task, learner, resampling, param_values = NULL, 
   }, set = names(sets), row_ids = sets)
   pdatas = discard(pdatas, is.null)
 
-  # In case we are parallelizing and the learner needs marshaling, we need to send the marshaled model back to
-  # the main process. We can skip this if we are using callr encapsulation as we already have the marshaled
-  # model.
-  marshal_maybe_after_predict(
-    learner = learner, store_models = store_models, is_sequential = is_sequential, model_marshaled = model_marshaled_or_null
+  # In case store_models is FALSE, we now discard the model
+  # Otherwise, we set the model to the marshaled model, if callr encapsulation was used and we are parallelizing
+  # If callr encapsulation was not used, we still have to marshal the model
+  set_model_after_predict(
+    learner = learner, store_models = store_models, is_sequential = is_sequential, model = model_marshaled_or_null
   )
 
   learner_state = set_class(learner$state, c("learner_state", "list"))
@@ -291,29 +289,34 @@ workhorse = function(iteration, task, learner, resampling, param_values = NULL, 
   list(learner_state = learner_state, prediction = pdatas, param_values = learner$param_set$values, learner_hash = learner_hash)
 }
 
-unmarshal_maybe_after_train = function(learner, store_models, is_sequential) {
-  if (!is_marshaled_model(learner$model)) return(NULL)
-  if (is_sequential) {
-    # no need to keep the marshaled model in RAM, because we can send back the unmarshaled model back to the main
-    # process
-    model_marshaled = NULL
-    learner$model = unmarshal_model(model = learner$model, inplace = TRUE)
-  } else {
-    # here, we simultaneously keep the marshaled and unmarshaled model in RAM
-    model_marshaled = learner$model
-    learner$model = unmarshal_model(model = learner$model, inplace = FALSE)
+get_marshaled_if_needed = function(learner, store_models, is_sequential, unmarshal) {
+  if (!is_marshaled_model(learner$model)) {
+    NULL
+  } else if (!store_models) { # else condition means model was marshaled
+    # not storing models, so no need to keep the marhshaled model
+    NULL
+  } else if (is_sequential && unmarshal) {
+    # because of no parallelization, we can send the unmarshaled model back, which the user
+    # wants anyway (because unmarshal = TRUE)
+    NULL
+  } else if (is_sequential && !unmarshal) {
+    # here, the user set unmarshal to FALSE, i.e. he wants the marshaled model, so we
+    # keep it, as we have already computed it
+    learner$model
+  } else { # parallel execution
+    # here, we just need to send back the marshaled model
+    learner$model
   }
-
-  model_marshaled
 }
 
-marshal_maybe_after_predict = function(learner, store_models, is_sequential, model_marshaled) {
+set_model_after_predict = function(learner, store_models, is_sequential, model) {
   if (!store_models) {
     lg$debug("Erasing stored model for learner '%s'", learner$id)
     learner$state$model = NULL
-  } else if (!is.null(model_marshaled)) {
+  } else if (!is.null(model)) {
     # callr + parallelization
-    learner$model = model_marshaled
+    # or callr + sequential, but unmarshal was FALSE
+    learner$model = model
   } else if (!is_sequential) {
     # parallelization without callr
     learner$model = marshal_model(learner$model, inplace = TRUE)