+++ NEWS.md
@@ -6,6 +6,7 @@
 * feat: dictionary conversion of `mlr_learners` respects prototype arguments
 recently added in mlr3misc
 * perf: skip unnecessary clone of learner's state in `resample()`
+* fix: Encapsulation methods use the same RNG state now.
 
 # mlr3 0.17.2