+++ NEWS.md
@@ -3,6 +3,7 @@
 * Feat: added support for `"marshal"` property, which allows learners to process
 models so they can be serialized. This happens automatically during `resample()`
 and `benchmark()`. The naming was inspired by the {marshal} package.
+* Log encapsulated errors and warnings with the `lgr` package.
 
 # mlr3 0.18.0
 
@@ -10,6 +11,7 @@ and `benchmark()`. The naming was inspired by the {marshal} package.
 * feat: dictionary conversion of `mlr_learners` respects prototype arguments
 recently added in mlr3misc
 * perf: skip unnecessary clone of learner's state in `resample()`
+* fix: Encapsulation methods use the same RNG state now.
 
 # mlr3 0.17.2