+++ NEWS.md
@@ -1,4 +1,4 @@
-# mlr3 (development version)
+# mlr3 0.19.0
 
 * Added support for `"marshal"` property, which allows learners to process models so they can be serialized.
 This happens automatically during `resample()` and `benchmark()`.